Adding filename parcing to diff header.
Adding regex and lazy static to facilitate easy use of the parcing
system for each line.

The regex used could be replaced without effecting much of the outer
system.

Seeing as anywhere else, filenames are repeated, using the header seems
like a good enough solution for the task that can be quickly exteneded.